plan component cleaned
